+++ sandbox/accounts/forms/__init__.py
@@ -1,0 +1,6 @@
+from .authentication import AuthenticationForm
+
+
+__all__ = [
+    "AuthenticationForm",
+]